+++ source/scan-tool/main.cpp
@@ -30,6 +30,7 @@
 #elif defined(_WIN32)
 #include <Windows.h>
 #endif
+#include "Handler7z.hpp"
 #include "HandlerAr.hpp"
 #include "HandlerCab.hpp"
 #include "HandlerGzip.hpp"
@@ -102,6 +103,8 @@ void showHelp()
             << "                                       VirusTotal, but does not get reports.\n"
             << "  --zip            - add ZIP file handler which extracts ZIP files and scans\n"
             << "                     each contained file, too.\n"
+            << "  --7zip | --7z    - add 7-Zip file handler which extracts 7-Zip files and\n"
+            << "                     scans each contained file, too.\n"
             << "  --tar            - add tape archive (*.tar) file handler which extracts tape\n"
             << "                     archives and scans each contained file, too.\n"
             << "  --gzip | --gz    - add gzip file handler which extracts gzip files and scans\n"
@@ -237,6 +240,7 @@ int main(int argc, char ** argv)
   //scan strategy
   scantool::virustotal::Strategy selectedStrategy = scantool::virustotal::Strategy::None;
   //flags for archive file handlers
+  bool handle7Zip = false;
   bool handleZIP = false;
   bool handleTar = false;
   bool handleGzip = false;
@@ -248,7 +252,7 @@ int main(int argc, char ** argv)
 
   if ((argc > 1) and (argv != nullptr))
   {
-    int i=1;
+    int i = 1;
     while (i < argc)
     {
       if (argv[i] != nullptr)
@@ -546,6 +550,17 @@ int main(int argc, char ** argv)
           }
           handleZIP = true;
         } //handle ZIP files
+        else if ((param == "--7zip") || (param == "--7z") || (param == "--7-zip"))
+        {
+          //Has the 7z option already been set?
+          if (handle7Zip)
+          {
+            std::cout << "Error: Parameter " << param << " must not occur more than once!"
+                      << std::endl;
+            return scantool::rcInvalidParameter;
+          }
+          handle7Zip = true;
+        } //handle 7-Zip files
         else if (param == "--tar")
         {
           //Has the tar option already been set?
@@ -780,6 +795,11 @@ int main(int argc, char ** argv)
   if (handleZIP)
   {
     strategy->addHandler(std::unique_ptr<scantool::virustotal::ZipHandler>(new scantool::virustotal::ZipHandler(ignoreExtractionErrors)));
+  }
+  //check, if user wants 7z handler
+  if (handle7Zip)
+  {
+    strategy->addHandler(std::unique_ptr<scantool::virustotal::Handler7z>(new scantool::virustotal::Handler7z(ignoreExtractionErrors)));
   }
   //check if user wants tar handler
   if (handleTar)